Extend the graphviz output format to support reverse flag
Earlier the graphviz output format didn't work if called with the
reverse flag. This is now fixed.

This commit also handles the case of more than one packages having
conflicting versions of the same dependency. In such cases, the edges
between dependency and the parent pkg will be labelled as per the
required version_spec of the respective dependency.