Implement 'ReversedPackageDAG.reverse' method
While this method is not being used in the script, it's been added for
completeness and to ensure that the APIs of all the classes involved
are clean.

Also, the 'DistPackage.as_required_by' method has been renamed to
'DistPackage.as_parent_of' for clarity.

Tests have been added and modified.